Clear session on guest auth failure
Problem:
Expiry of the guest token is encoded on both the server and the client.
Although the digits client detects when a guest token expires as per its
semantics and refreshes it, it does not _react_ when an api response
indicates that a guest token has expired on the server.

Solution:
In this change, we clear the guest session when any api interaction
indicates an expired guest session.

Change-Id: I71c3f50c7dc901ee5917f6dcf271afb4afd93790