OkHttp 2.3.1+ expects a body in post requests.
Problem:
OkHttp 2.3.1+ expects a body in post requests. We have previously
converted our @Query params to @Fields to work around this problem. See:
https://gerrit.twitter.biz/#/c/98776/

In this change we pass an empty body to the delete contacts api
interface provided to retroft. This is the recommended workaround
provided by the author of Okhttp. See:
https://github.com/square/retrofit/issues/854

Change-Id: I14f90feeb1845371a20adbb8356f035d830a3de2